requirements.txt
anyio==4.4.0
certifi==2024.8.30
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
python-dotenv==1.0.1
python-telegram-bot==21.5
sniffio==1.3.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
sqlalchemy-libsql